tender-doc-filler/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-docx==1.1.2
python-multipart==0.0.9
anthropic==0.40.0
pydantic==2.9.0
pytest==8.3.0
httpx==0.27.0
